services/pcap-processor/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scapy==2.5.0
httpx==0.27.0
python-dotenv==1.0.1
pyshark==0.6
